Update main.tex
Mining is viable for entry-level users, but it is not appealing in its usual form (e.g. bitcoin mining). Creating a mining application is not *very* difficult. (It's probably easy to implement, but using money for everything is just easier.). A solution is not 'given', but 'proposed'. Yeah, maybe mining is easier.